+++ vignettes_src/hintr-example.R
@@ -55,8 +55,7 @@ hintr_options <- list(
   output_aware_plhiv = "true",
   rng_seed = 17,
   no_of_samples = 20,
-  max_iter = 250,
-  permissive = "false"
+  max_iter = 250
 )
 
 calibration_options <- list(